Fix popover content might exceed container boundary on Safari & Chrome
It seems to be how browser determines layer stacking.

In <Popover> with a list of <Checkbox>, the checkbox icon and the aside label has their own
stacking context, therefore they are not clipped by the .popover__container in Safari and Chrome.
However, Firefox does clip those content.

This commit tries to fix this issue by force-creating a new stacking context on the popover container,
so the checkbox icon and aside label are now nested inside the context.