Force publish all packages when releasing canary
This restores the v2 behavior.
When we were auto-releasing a canary build for #220, we found it only
published a new build for '@ichef/gypctete-imageeditor'.

Unfortunately the version of '@ichef/gypctete' listed in that build
remains "^3.0.0" because the two packages were not published at the
same time.

By force-publish packages all together, we can at least be sure their
dependencies would be correct.