Fix anchored popover top position when document is scrollable
Anchored popover is a <Popover> wrapped with two HOCs:
`closable()(anchored()(Popover)`

Before this commit it will form following DOM structure:

```html
<body>
  <div class="gyp-base-layer">
    <div class="gyp-closable>
      <div class="gyp-popover gyp-popover--top">
    </div>
  </div>
</body>
```

`gyp-popover` is absolutely positioned, with `top` value which
is relative to its document top left corner. `gyp-closable` is
`position: fixed` with `width: 100vw` and `height: 100vh`.

It works when document is too short to scroll, such as previous
storybook doc canvas. But when the element triggering popover is
over the screen(i.e. we have to scroll to click it), the position
of anchored popover will be wrong because the `position: absoulte`
of `gyp-popover` is relative to `gyp-closable`, which is
`position: fixed`, so the `top` value is wrong.

The position of `gyp-popover` should always be relative to the
document. So in this fix we change the `closable` HOC code to change
DOM structure as:

```html
<body>
  <div class="gyp-base-layer">
    <div class="gyp-closable" />
    <div class="gyp-popover" />
  </div>
</body>
```

Let the `gyp-popover` be relative to `gyp-base-layer`, which is absolute
to `document.body`.

Note that this bug won't happen on Cloud2 since on Cloud2 the
`document.body` height is always `100vh`, which is not scrollable.